Move menus to components/menus
